BEEEP: Move middleware to maintain readability.
